Moved the method LocationUtils#getNearbyBlocks to the new class BlockUtils
